orders_api/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
requests
httpx
confluent-kafka
python-keycloak